[Refactoring] Merge into slice.test.js
- Delete actions.test.js and Merge into slice.test.js
- Add test async actions test
